requirements.txt
PyYAML 
recursive-diff
boto3
retrying
pymongo
dnspython
requests
